code fixed

+++ app/Http/Controllers/Api/TasksController.php
@@ -15,21 +15,16 @@ use Illuminate\Support\Str;
 
 class TasksController extends BaseController
 {
-    public function tasks(): \Illuminate\Contracts\View\Factory|\Illuminate\Foundation\Application|\Illuminate\Contracts\View\View|\Illuminate\Contracts\Foundation\Application
-    {
-        $tasks = Tasks::all();
-        return view('tasks',compact('tasks'));
-    }
     /**
-     * List
-     * @param  array
+     * Table
+     * @param array
      * @return JsonResponse
      */
 
     public function index(): JsonResponse
     {
-        $tasks = Tasks::all();
-        return $this->sendResponse($tasks,"Barcha tasklar yuborildi");
+        $tasks = Tasks::with('sample')->get();
+        return $this->sendResponse($tasks, "Barcha masalalar ro'yhati");
     }
 
     /**
@@ -40,7 +35,7 @@ class TasksController extends BaseController
      */
     public function store(Request $request): \Illuminate\Http\JsonResponse
     {
-        $request->validate([
+        $data = $request->validate([
             'name' => 'nullable|string|max:255',
             'condition' => 'nullable|string',
             'input' => 'nullable|string',
@@ -50,22 +45,28 @@ class TasksController extends BaseController
             'memory' => 'required|integer|min:0',
             'difficulty' => 'required|integer|min:0',
             'user_id' => 'required|exists:users,id',
+            'sample_test' => 'nullable|array',
+            'sample_test.*.input' => 'required|string',
+            'sample_test.*.output' => 'required|string',
         ]);
 
         $data = $request->all();
+
         $data['key'] = Str::uuid();
+
         $task = Tasks::create($data);
 
-        foreach ($request->get('sample_test') as $sample) {
-            TaskInputOutput::create([
-                "task_id"   => $task->id,
-                "input"     => $sample['input'] ?? "",
-                "output"    => $sample['output'] ?? ""
-            ]);
+
+        if ($request->get('sample_test')) {
+            foreach ($request->get('sample_test') as $sample) {
+                TaskInputOutput::create([
+                    "task_id" => $task->id,
+                    "input" => $sample['input'] ?? "",
+                    "output" => $sample['output'] ?? ""
+                ]);
+            }
+            $task['sample'] = $task->sample;
         }
-
-
-        $task['sample'] = $task->sample;
 
         return $this->sendResponse(array($task), "Task muvaffaqiyatli yaratildi.");
     }
@@ -92,7 +93,7 @@ class TasksController extends BaseController
      * @return JsonResponse
      */
 
-    public function update(Request $request,int $id): \Illuminate\Http\JsonResponse
+    public function update(Request $request, int $id): \Illuminate\Http\JsonResponse
     {
         $request->validate([
             'name' => 'nullable|string|max:255',
@@ -104,6 +105,9 @@ class TasksController extends BaseController
             'memory' => 'required|integer|min:0',
             'difficulty' => 'required|integer|min:0',
             'user_id' => 'required|exists:users,id',
+            'sample_test' => 'nullable|array',
+            'sample_test.*.input' => 'required|string',
+            'sample_test.*.output' => 'required|string',
         ]);
 
         $data = $request->all();
@@ -111,23 +115,22 @@ class TasksController extends BaseController
         $task = Tasks::find($id);
 
         if ($task) {
-            TaskInputOutput::where('task_id',$task->id)->delete();
+            if ($request->get('sample_test')) {
+                TaskInputOutput::where('task_id', $task->id)->delete();
 
-            foreach ($request->get('sample_test') as $sample) {
-                TaskInputOutput::create([
-                    "task_id"   => $task->id,
-                    "input"     => $sample['input'] ?? "",
-                    "output"    => $sample['output'] ?? ""
-                ]);
+                foreach ($request->get('sample_test') as $sample) {
+                    TaskInputOutput::create([
+                        "task_id" => $task->id,
+                        "input" => $sample['input'] ?? "",
+                        "output" => $sample['output'] ?? ""
+                    ]);
+                }
+
+                $task->update($data);
+                $task['sample'] = $task->sample;
             }
-
-            $task->update($data);
-            $task['sample'] = $task->sample;
-
             return $this->sendResponse(array($task), "Task muvaffaqiyatli yangilandi.");
-        }
-
-        else {
+        } else {
             return $this->sendError("Bunday masala mavjud emas");
         }
     }
@@ -143,7 +146,7 @@ class TasksController extends BaseController
         $task = Tasks::find($id);
         if ($task) {
             $task->delete();
-            return $this->sendResponse([],"Task o'chirildi.");
+            return $this->sendResponse([], "Task o'chirildi.");
         }
         return $this->sendError("Bunday masala saqlanmagan");
     }
@@ -153,6 +156,8 @@ class TasksController extends BaseController
      *
      * @param Request $request
      * @param int $task_id
+     * @requestMediaType multipart/form-data
+
      * @throws Exception
      */
     public function storeTestFile(Request $request, int $task_id)
@@ -181,7 +186,7 @@ class TasksController extends BaseController
      */
     public function updateTestFile(Request $request, int $task_id)
     {
-        $test = Test::where('task_id',$task_id)->first();
+        $test = Test::where('task_id', $task_id)->first();
 
         $file = $request->file('test_file');
         $newFilePath = $file->store('tests');
@@ -205,7 +210,7 @@ class TasksController extends BaseController
     {
         $zip = new \ZipArchive;
         if ($zip->open(storage_path('app/' . $filePath)) === TRUE) {
-            $extractPath = storage_path('app/tests/'.$test->id);
+            $extractPath = storage_path('app/tests/' . $test->id);
             $zip->extractTo($extractPath);
             $zip->close();
